SOme changes in repositories

+++ src/main/java/inz/repository/PromotionRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.math.BigInteger;
 
 @Repository
-public interface PromotionRepository extends JpaRepository<Promotion, BigInteger> {
+public interface PromotionRepository extends JpaRepository<Promotion, Integer> {
 
 }
