+++ src/main/java/inz/repository/UserRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.math.BigInteger;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, BigInteger>, UserRepositoryCustom {
+public interface UserRepository extends JpaRepository<User, Integer>, UserRepositoryCustom {
 
 }
